test train supervised rf, logistic
